compilation fix on one 64 (size_t!=unsigned int)

+++ src/tools/replace_text.cc
@@ -137,7 +137,7 @@ main(int argc, char ** argv)
 
 
 		// sane values
-		to = std::min(to, pdf->getPageCount()+1);
+		to = std::min(static_cast<unsigned int>(to), pdf->getPageCount()+1);
 
 		// now the hard stuff comes - do this crazy loops intentionally
 		for (size_t things_to_replace = 0; things_to_replace < withs.size(); ++things_to_replace)
